fix: 주엽님과의 작업 사항 합치기
